refactor(core-api): update pie chart data handling by adding new dependencies and consolidating imports for improved clarity
