+++ packages/core-api/package.json
@@ -27,6 +27,7 @@
   },
   "devDependencies": {
     "@stylistic/eslint-plugin": "^2.6.1",
+    "@types/d3-shape": "^3.1.6",
     "@types/eslint": "^8.56.11",
     "@types/node": "^20.17.9",
     "@typescript-eslint/eslint-plugin": "^8.0.0",
@@ -44,5 +45,8 @@
     "rimraf": "^6.0.1",
     "typescript": "^5.6.3",
     "vitest": "^2.1.9"
+  },
+  "dependencies": {
+    "d3-shape": "^3.2.0"
   }
 }
