+++ packages/plugin-awesome/src/generators.ts
@@ -6,6 +6,7 @@ import {
   incrementStatistic,
   nullsLast,
   ordinal,
+  getPieChartValues,
 } from "@allurereport/core-api";
 import {
   type AllureStore,
@@ -32,7 +33,6 @@ import {
   createReportDataScript,
   createScriptTag,
   createStylesLinkTag,
-  getPieChartValues,
 } from "@allurereport/web-commons";
 import Handlebars from "handlebars";
 import { readFile } from "node:fs/promises";
