+++ packages/plugin-awesome/src/model.ts
@@ -1,4 +1,5 @@
-import type { CiDescriptor, ChartOptions, EnvironmentsConfig, TestResult } from "@allurereport/core-api";
+import type { CiDescriptor, EnvironmentsConfig, TestResult } from "@allurereport/core-api";
+import type { ChartOptions } from "@allurereport/plugin-api";
 
 export type AwesomeOptions = {
   reportName?: string;
