+++ packages/plugin-classic/src/generators.ts
@@ -5,6 +5,7 @@ import {
   type TreeGroup,
   type TreeLeaf,
   compareBy,
+  getPieChartValues,
   incrementStatistic,
   nullsLast,
   ordinal,
@@ -32,7 +33,6 @@ import {
   createReportDataScript,
   createScriptTag,
   createStylesLinkTag,
-  getPieChartValues,
 } from "@allurereport/web-commons";
 import Handlebars from "handlebars";
 import { readFile } from "node:fs/promises";
