+++ packages/web-awesome/src/stores/chart.ts
@@ -1,8 +1,8 @@
-import { fetchReportJsonData, getPieChartValues, createCharts } from "@allurereport/web-commons";
+import { fetchReportJsonData, createCharts } from "@allurereport/web-commons";
 import { signal } from "@preact/signals";
 import type { StoreSignalState } from "@/stores/types";
-import type { ChartsResponse, UIChartsData, PieChartValues } from "@allurereport/web-commons";
-import type { Statistic } from "@allurereport/core-api";
+import type { ChartsResponse, UIChartsData } from "@allurereport/web-commons";
+import { type PieChartValues, type Statistic, getPieChartValues } from "@allurereport/core-api";
 
 export const pieChartStore = signal<StoreSignalState<PieChartValues>>({
   loading: true,
