+++ packages/web-commons/package.json
@@ -27,14 +27,12 @@
   },
   "devDependencies": {
     "@stylistic/eslint-plugin": "^2.6.1",
-    "@types/d3-shape": "^3.1.6",
     "@types/eslint": "^8.56.11",
     "@typescript-eslint/eslint-plugin": "^8.0.0",
     "@typescript-eslint/parser": "^8.0.0",
     "@vitest/runner": "^2.1.9",
     "allure-js-commons": "^3.3.0",
     "allure-vitest": "^3.3.0",
-    "d3-shape": "^3.2.0",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
