--- packages/web-commons/src/charts.ts
@@ -1,23 +1,0 @@
-import type {
-  BaseTrendSliceMetadata,
-  TrendSlice,
-  TrendSliceMetadata,
-  BasePieSlice
-} from "@allurereport/core-api";
-import type { PieArcDatum } from "d3-shape";
-import { arc, pie } from "d3-shape";
-
-export const d3Arc = arc<PieArcDatum<BasePieSlice>>().innerRadius(40).outerRadius(50).cornerRadius(2).padAngle(0.03);
-
-export const d3Pie = pie<BasePieSlice>()
-  .value((d) => d.count)
-  .padAngle(0.03)
-  .sortValues((a, b) => a - b);
-
-export interface StatusMetadata extends BaseTrendSliceMetadata {}
-export type StatusTrendSliceMetadata = TrendSliceMetadata<StatusMetadata>;
-export type StatusTrendSlice = TrendSlice<StatusTrendSliceMetadata>;
-
-export interface SeverityMetadata extends BaseTrendSliceMetadata {}
-export type SeverityTrendSliceMetadata = TrendSliceMetadata<SeverityMetadata>;
-export type SeverityTrendSlice = TrendSlice<SeverityTrendSliceMetadata>;
