+++ packages/web-commons/src/index.ts
@@ -3,7 +3,5 @@ export * from "./static.js";
 export * from "./attachments.js";
 export * from "./i18n.js";
 export * from "./strings.js";
-export * from "./charts.js";
 export * from "./charts/colors.js";
 export * from "./charts/utils.js";
-export * from "./charts/getPieChart.js";
