+++ packages/web-components/src/components/Charts/SuccessRatePieChart/index.tsx
@@ -1,5 +1,4 @@
-import { type TestStatus } from "@allurereport/core-api";
-import type { PieChartValues } from "@allurereport/web-commons";
+import type { PieChartValues, TestStatus } from "@allurereport/core-api";
 import cx from "clsx";
 import { Heading } from "@/components/Typography";
 import styles from "./styles.scss";
