+++ packages/web-summary/src/components/ReportCard/index.tsx
@@ -1,5 +1,5 @@
-import { type Statistic, type TestStatus, formatDuration } from "@allurereport/core-api";
-import { capitalize, getPieChartValues } from "@allurereport/web-commons";
+import { type Statistic, type TestStatus, formatDuration, getPieChartValues } from "@allurereport/core-api";
+import { capitalize } from "@allurereport/web-commons";
 import { Heading, StatusLabel, SuccessRatePieChart, Text, allureIcons } from "@allurereport/web-components";
 import type { FunctionalComponent } from "preact";
 import IconLabel from "@/components/IconLabel";
